performance foundational work for suppression effect
