+++ @twc_framework/addons/twc_suppress/CfgEventHandlers.hpp
@@ -1,0 +1,7 @@
+class Extended_FiredBIS_EventHandlers {
+	class AllVehicles {
+		class TWC_Suppress {
+			clientFiredBIS = "_this call twc_suppress_fnc_fired";
+		};
+	};
+};
